refactor(api): extract domain types to internal/types
Move pkg/api data types and their methods into a new internal package
pkg/api/internal/types. pkg/api keeps re-exports (type X = types.X,
const/var re-exports) so callers stay unchanged. The re-exports let us
minimize the size of the change and avoid updating any tests in this
commit. Eventually, after splitting into the 3 packages (dbreads, admin,
runtime), we will drop the re-exports in the api package and callers
can reference the types package directly.

Prerequisite for extracting the read layer into pkg/api/internal/dbreads,
which is itself a step towards later splitting the admin and runtime
writer packages. dbreads cannot import pkg/api without an import cycle,
so shared types have to live in a neutral internal package first.

This commit is purely refactoring and shouldn't change any behavior.

Refs: #1375
