+++ backend/pkg/api/actions.go
@@ -1,26 +1,12 @@
 package api
 
 import (
-	"time"
-
 	"github.com/doug-martin/goqu/v9"
+
+	"github.com/flatcar/nebraska/backend/pkg/api/internal/types"
 )
 
-// FlatcarAction represents an Omaha action with some Flatcar specific fields.
-type FlatcarAction struct {
-	ID                    string    `db:"id" json:"id"`
-	Event                 string    `db:"event" json:"event"`
-	ChromeOSVersion       string    `db:"chromeos_version" json:"chromeos_version"`
-	Sha256                string    `db:"sha256" json:"sha256"`
-	NeedsAdmin            bool      `db:"needs_admin" json:"needs_admin"`
-	IsDelta               bool      `db:"is_delta" json:"is_delta"`
-	DisablePayloadBackoff bool      `db:"disable_payload_backoff" json:"disable_payload_backoff"`
-	MetadataSignatureRsa  string    `db:"metadata_signature_rsa" json:"metadata_signature_rsa"`
-	MetadataSize          string    `db:"metadata_size" json:"metadata_size"`
-	Deadline              string    `db:"deadline" json:"deadline"`
-	CreatedTs             time.Time `db:"created_ts" json:"created_ts"`
-	PackageID             string    `db:"package_id" json:"-"`
-}
+type FlatcarAction = types.FlatcarAction
 
 // AddFlatcarAction registers the provided Omaha Flatcar action.
 func (api *API) AddFlatcarAction(action *FlatcarAction) (*FlatcarAction, error) {
