+++ backend/pkg/api/channels.go
@@ -3,10 +3,10 @@ package api
 import (
 	"database/sql"
 	"errors"
-	"time"
 
 	"github.com/doug-martin/goqu/v9"
-	"gopkg.in/guregu/null.v4"
+
+	"github.com/flatcar/nebraska/backend/pkg/api/internal/types"
 )
 
 var (
@@ -19,17 +19,7 @@ var (
 	ErrBlacklistedChannel = errors.New("nebraska: blacklisted channel")
 )
 
-// Channel represents a Nebraska application's channel.
-type Channel struct {
-	ID            string      `db:"id" json:"id"`
-	Name          string      `db:"name" json:"name"`
-	Color         string      `db:"color" json:"color"`
-	CreatedTs     time.Time   `db:"created_ts" json:"created_ts"`
-	ApplicationID string      `db:"application_id" json:"application_id"`
-	PackageID     null.String `db:"package_id" json:"package_id"`
-	Package       *Package    `db:"package" json:"package"`
-	Arch          Arch        `db:"arch" json:"arch"`
-}
+type Channel = types.Channel
 
 // AddChannel registers the provided channel.
 func (api *API) AddChannel(channel *Channel) (*Channel, error) {
